Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/WebAssembly-allowed-expected.txt

CONSOLE MESSAGE: line 13: PASS
